requirements.txt.txt
streamlit
pandas
numpy
matplotlib
fpdf2